Commit dynamic content to master

+++ deploy.sh
@@ -5,19 +5,24 @@ if [ $TRAVIS_PULL_REQUEST == "true" ]; then
 fi
 
 set -e
-
-rm -rf _site
-mkdir _site
-
-git clone https://${GITHUB_TOKEN}@github.com/KovuTheHusky/kovuthehusky.com.git --branch gh-pages _site
-
-yarn
-bundle exec jekyll build
-
-cd _site
-ruby places.rb ${FOURSQUARE_TOKEN}
 git config user.email "kovuthehusky@gmail.com"
 git config user.name "KovuTheHusky"
-git add --all
-git commit -a -m "Travis #$TRAVIS_BUILD_NUMBER"
-git push --force origin gh-pages
+
+if [ $TRAVIS_EVENT_TYPE == "cron" ]; then
+    ruby places.rb ${FOURSQUARE_TOKEN}
+    ruby pokemongo.rb
+    ruby projects.rb
+    git add --all
+    git commit -a -m "Travis #$TRAVIS_BUILD_NUMBER"
+    git push origin master
+else
+    rm -rf _site
+    mkdir _site
+    git clone https://${GITHUB_TOKEN}@github.com/KovuTheHusky/kovuthehusky.com.git --branch gh-pages _site
+    yarn
+    bundle exec jekyll build
+    cd _site
+    git add --all
+    git commit -a -m "Travis #$TRAVIS_BUILD_NUMBER"
+    git push --force origin gh-pages
+fi
